Error handling is done as a Middleware function

+++ Backend/dataaccesslayer/userdal.js
@@ -1,8 +1,9 @@
 const dbcon = require("../libs/dbcon");
+const mysql = require('mysql');
+const ApiError = require('../middleware/error/apierror')
+const getLowerCaseKeys = require('../libs/commonclass');
+const apiErrorHandler = require("../middleware/Error/errorhandling");
 const User = require("../model/user");
-const bodyParser = require('body-parser')
-const mysql = require('mysql')
-
 
 
 class userDAL {
@@ -10,20 +11,33 @@ class userDAL {
     getAllUsers(filters) {
         return new Promise((resolve, reject) => {
             let sqlQuery = "SELECT * FROM tbl_User";
-            if (Object.keys(filters).length != 0) {
+            
+            if(Object.keys(filters).length != 0 &&
+            ((filters.limit == undefined && filters.skip != undefined) ||
+             (filters.sort == undefined && filters.order != undefined))){
+                return reject(ApiError.improperRequest("Improper Request Error: You can't query OFFSET " 
+                +"without LIMIT or you can't order without sort: getAllUser::UserDAL.js"));
+            }
+            if (Object.keys(filters).length != 0 &&
+                (filters.limit == undefined) && (filters.sort == undefined)) {
                 sqlQuery = sqlQuery + "\n" + "WHERE "
                 for (var key in filters) {
                     if ((key != "limit") && (key != "skip") && (key != "sort") && (key != "order")) {
                         key = (mysql.escape(key))
                         key = key.replace("'", "");
                         key = key.replace("'", "");
-                        sqlQuery = sqlQuery + key + " = " + (mysql.escape((filters[key])))
+                        sqlQuery = sqlQuery + "\n" + key + " = " + (mysql.escape((filters[key])))
                     }
                 }
             }
             if (Object.keys(filters).length != 0 && (filters.sort != undefined)) {
+                filters = getLowerCaseKeys(filters)
                 sqlQuery = sqlQuery + "\n" + "ORDER BY "
-                var order = (filters.order == undefined) ? 1 : filters.order
+                var order = (filters.order != undefined) ? parseInt(filters.order) : undefined;
+                if (order != undefined && isNaN(order)) {
+                    return reject(ApiError.improperRequest("Improper Request Error: Order must be integer getAllUser::UserDAL.js"))
+                }
+                order = (filters.order == undefined) ? 1 : filters.order
                 order = (filters.order == 1) ? "ASC" : "DESC"
                 order = mysql.escape(order)
                 order = order.replace("'", ' ');
@@ -31,38 +45,42 @@ class userDAL {
                 var sort = mysql.escape(filters.sort)
                 sort = sort.replace("'", ' ');
                 sort = sort.replace("'", ' ');
-                sqlQuery = sqlQuery + sort + " " + order 
+                sqlQuery = sqlQuery + sort + " " + order
             }
             if (filters != null && (filters.limit != undefined)) {
-                var limit = (", "+ (filters.limit))
-                var skip = (filters.skip);
-                if (skip == undefined){
-                    skip =  "";
+                var skip = (filters.skip != undefined) ? parseInt(filters.skip) : undefined;
+                var limit = (parseInt(filters.limit))
+                if ((skip != undefined && isNaN(skip)) || isNaN(limit)) {
+                    return reject(ApiError.improperRequest("Improper Request Error: Skip or Limit must be integer: getAllUser::UserDAL.js:"))
+                }
+                limit = (", " + (filters.limit))
+                skip = (filters.skip);
+                if (skip == undefined) {
+                    skip = "";
                     limit = ((parseInt(filters.limit)));
                 }
                 skip = mysql.escape(skip).replace("'", " ");
                 skip = skip.replace("'", " ");
                 limit = mysql.escape(limit).replace("'", "");
                 limit = limit.replace("'", "");
-                sqlQuery = sqlQuery + "\n" + "LIMIT " + skip + limit+ "\n" ;
+                sqlQuery = sqlQuery + "\n" + "LIMIT " + skip + limit + "\n";
             }
-            
+
             dbcon.getNewConnection(function (err, connection) {
                 if (err) {
-                    console.log("An error occured while connecting db");
-                    reject(err);
+                    return reject(ApiError.databaseConnectionError("Database Connection Error: getAllUser::UserDAL.js: " + err.message));
                 } else {
                     console.log("Database connection succesfully")
+                    console.log(sqlQuery);
                     connection.query(sqlQuery, (err, results) => {
                         if (err) {
-                            console.log("An Error Occured GetAllUsers::UserDAL.js " + err.message);
                             dbcon.closeConnection(connection);
-                            reject(err);
+                            return reject(ApiError.databaseError("Database Error: getAllUser::UserDAL.js:" + err.message));
                         }
                         else {
                             console.log("Succesfully");
                             dbcon.closeConnection(connection);
-                            resolve(results);
+                            return resolve(results);
                         }
                     })
                 }
@@ -74,31 +92,29 @@ class userDAL {
 
 
     getUser(id) {
-        return new Promise((resolve, reject) => {
+        
             let sqlQuery = "SELECT * FROM tbl_User Where ? ";
             dbcon.getNewConnection(function (err, connection) {
                 if (err) {
-                    console.log("An error occured while connecting db");
-                    reject(err)
+                    return (ApiError.databaseConnectionError("Database Connection Error: getUser::UserDAL.js:" + err.message));
                 }
                 else {
                     console.log("Database connection succesfully")
                     var condition = { UserID: id }
                     connection.query(sqlQuery, [condition], (err, results) => {
                         if (err) {
-                            console.log("An Error Occured getUser::UserDAL.js " + err.message);
                             dbcon.closeConnection(connection);
-                            reject(err);
+                            return (ApiError.databaseError("Database Error: getUser::UserDAL.js:" + err.message));
                         }
                         else {
                             console.log("Succesfully");
                             dbcon.closeConnection(connection);
-                            resolve(results);
+                            return (results);
                         }
                     });
                 }
             })
-        })
+        
     }
 
     createUser(user) {
@@ -106,8 +122,7 @@ class userDAL {
             let sqlQuery = "INSERT INTO tbl_User(Name, Surname, Email, PhoneNo) VALUES ?"
             dbcon.getNewConnection(function (err, connection) {
                 if (err) {
-                    console.log("An error occured while connecting db");
-                    reject(err)
+                    return reject(ApiError.databaseConnectionError("Database Connection Error: createUser::UserDAL.js:" + err.message));
                 }
                 else {
                     console.log("Database connection succesfully")
@@ -116,14 +131,13 @@ class userDAL {
                     ];
                     connection.query(sqlQuery, [Values], (err) => {
                         if (err) {
-                            console.log("An Error Occured CreateUser::UserDAL.js " + err.message);
                             dbcon.closeConnection(connection);
-                            reject(err);
+                            return reject(ApiError.databaseError("Database Error:  createUser::UserDAL.js: " + err.message));
                         }
                         else {
                             console.log("Succesfully");
                             dbcon.closeConnection(connection);
-                            resolve(true);
+                            return resolve(true);
                         }
                     });
                 }
@@ -136,26 +150,28 @@ class userDAL {
 
     updateUser(user) {
         return new Promise((resolve, reject) => {
+            user = this.getUser(user.id)
+            if(user==null){
+                return reject(ApiError.unknownUser("User Not Found Error:updateUser::UserDAL.js"))
+            }
             var userForUpdate = { Name: user.name, Surname: user.surname, Email: user.email, PhoneNo: user.phoneNo }
             var condition = { UserID: user.id }
             let sqlQuery = "UPDATE tbl_User SET ? WHERE ?";
             dbcon.getNewConnection(function (err, connection) {
                 if (err) {
-                    console.log("An error occured while connecting db");
-                    reject(err);
+                    return reject(ApiError.databaseConnectionError("Database Connection Error: updateUser::UserDAL.js: " + err.message));
                 }
                 else {
                     console.log("Database connection succesfully")
                     connection.query(sqlQuery, [userForUpdate, condition], (err) => {
                         if (err) {
-                            console.log("An Error Occured CreateUser::User.js " + err.message);
                             dbcon.closeConnection(connection);
-                            reject(err)
+                            return reject(ApiError.databaseError("Database Error: updateUser::UserDAL.js: " + err.message));
                         }
                         else {
                             console.log("Succesfully");
                             dbcon.closeConnection(connection);
-                            resolve(true);
+                            return resolve(true);
                         }
                     });
                 }
@@ -166,26 +182,30 @@ class userDAL {
 
     deleteUser(condition) {
         return new Promise((resolve, reject) => {
-            var conditionForDelete = { "UserID": condition }
-            let sqlQuery = "DELETE FROM tbl_User WHERE ?"
+            var user = new User();
+            user = this.getUser(condition)
+            
+            if(user==null){
+                return reject(ApiError.unknownUser("User Not Found Error:deleteUser::UserDAL.js"))
+            }
+            condition = mysql.escape(condition)
+
+            let sqlQuery = "DELETE FROM tbl_User WHERE UserID = " + condition 
             dbcon.getNewConnection(function (err, connection) {
                 if (err) {
-                    console.log("An error occured while connecting db");
-                    reject(err)
-
+                    return reject(ApiError.databaseConnectionError("Database Connection Error: deleteUser::UserDAL.js: " + err.message));
                 }
                 else {
                     console.log("Database connection succesfully")
-                    connection.query(sqlQuery, conditionForDelete, (err) => {
+                    connection.query(sqlQuery, condition, (err) => {
                         if (err) {
-                            console.log("An Error Occured deleteUser::UserDAL.js " + err.message);
                             dbcon.closeConnection(connection);
-                            reject(err);
+                            return reject(ApiError.databaseError("Database Error: deleteUser::UserDAL.js: " + err.message));
                         }
                         else {
                             console.log("Succesfully");
                             dbcon.closeConnection(connection);
-                            resolve(true)
+                            return resolve(true)
                         }
                     })
                 }
